+++ test/store/test_store.py
@@ -1,3 +1,4 @@
+import tempfile
 import uuid
 from pathlib import Path
 from unittest.mock import patch
@@ -6,7 +7,7 @@ import minio.datatypes
 import pytest
 
 from climatoology.base.artifact import ArtifactModality, _Artifact
-from climatoology.store.object_store import MinioStorage
+from climatoology.store.object_store import MinioStorage, DataGroup
 
 
 @pytest.fixture
@@ -24,7 +25,8 @@ def mocked_client():
 
 
 def test_minio_save(mocked_client, general_uuid, default_artifact):
-    store_id = mocked_client['minio_storage'].save(default_artifact)
+    with patch.object(tempfile._RandomNameSequence, attribute='characters', new='a') as _:
+        store_id = mocked_client['minio_storage'].save(default_artifact)
 
     mocked_client['minio_client'].assert_called_once_with(
         endpoint='minio.test.org:9999',
@@ -32,20 +34,26 @@ def test_minio_save(mocked_client, general_uuid, default_artifact):
         secret_key='secret',
         secure=True,
     )
-    mocked_client['minio_client']().fput_object.assert_called_once_with(
+    mocked_client['minio_client']().fput_object.assert_any_call(
         bucket_name='test_bucket',
         object_name=f'{general_uuid}/{store_id}',
         file_path=str(default_artifact.file_path),
         metadata={
-            'Name': 'test_name',
-            'Modality': 'MAP_LAYER_GEOJSON',
-            'Original-Filename': default_artifact.file_path.name,
-            'Summary': 'Test summary',
-            'Description': 'Test description',
-            'Correlation-UUID': general_uuid,
-            'Store-ID': store_id,
+            'Type': DataGroup.DATA,
+            'Metadata-Object-Name': f'{general_uuid}/{store_id}.metadata.json',
         },
     )
+
+    mocked_client['minio_client']().fput_object.assert_any_call(
+        bucket_name='test_bucket',
+        object_name=f'{general_uuid}/{store_id}.metadata.json',
+        file_path='/tmp/tmpaaaaaaaa',
+        metadata={
+            'Type': DataGroup.METADATA,
+            'Data-Object-Name': f'{general_uuid}/{store_id}',
+        },
+    )
+    assert mocked_client['minio_client']().fput_object.call_count == 2
 
 
 def test_minio_save_all(mocked_client, general_uuid, default_artifact):
@@ -59,24 +67,30 @@ def test_minio_save_all(mocked_client, general_uuid, default_artifact):
         correlation_uuid=second_correlation_uuid,
     )
     mocked_client['minio_storage'].save_all([default_artifact, second_plugin_artifact])
-    assert mocked_client['minio_client']().fput_object.call_count == 2
+    assert mocked_client['minio_client']().fput_object.call_count == 4
 
 
 def test_minio_list_all(mocked_client, general_uuid, default_artifact):
-    return_mock = minio.datatypes.Object(
-        bucket_name='test_bucket',
-        object_name=f'{general_uuid}/{general_uuid}',
-        metadata={
-            'X-Amz-Meta-Name': 'test_name',
-            'X-Amz-Meta-Modality': 'MAP_LAYER_GEOJSON',
-            'X-Amz-Meta-Original-Filename': str(default_artifact.file_path),
-            'X-Amz-Meta-Summary': 'Test summary',
-            'X-Amz-Meta-Description': 'Test description',
-            'X-Amz-Meta-Store-Id': f'{general_uuid}_test_file.tiff',
-            'X-Amz-Meta-Params': '{"test param key": "test param val"}',
-        },
-    )
-    mocked_client['minio_client']().list_objects.return_value = iter([return_mock])
+    return_mock = [
+        minio.datatypes.Object(
+            bucket_name='test_bucket',
+            object_name=f'{general_uuid}/{general_uuid}',
+            metadata={
+                'X-Amz-Meta-Type': DataGroup.DATA.value,
+                'X-Amz-Meta-Metadata-Object-Name': f'{general_uuid}/{general_uuid}.metadata.json',
+            },
+        ),
+        minio.datatypes.Object(
+            bucket_name='test_bucket',
+            object_name=f'{general_uuid}/{general_uuid}.metadata.json',
+            metadata={
+                'X-Amz-Meta-Type': DataGroup.METADATA.value,
+                'X-Amz-Meta-Data-Object-Name': f'{general_uuid}/{general_uuid}',
+            },
+        ),
+    ]
+    mocked_client['minio_client']().list_objects.return_value = iter(return_mock)
+    mocked_client['minio_client']().get_object.return_value.json.return_value = default_artifact.model_dump(mode='json')
 
     result = mocked_client['minio_storage'].list_all(general_uuid)
     assert result == [default_artifact]
@@ -95,5 +109,5 @@ def test_minio_fetch(mocked_client, general_uuid):
     mocked_client['minio_client']().fget_object.assert_called_once_with(
         bucket_name='test_bucket',
         object_name=f'{general_uuid}/{store_id}',
-        file_path=Path('/tmp') / store_id,
+        file_path=f'/tmp/{store_id}',
     )